[Refactor] 폴더명 1차 변경
