+++ src/components/sdummy/dashboard.ts
