+++ src/pages/dashboard/[dashboardId]/edit.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import ChangeBebridge from "@/components/modal/ChangeBebridge";
-import HeaderBebridge from "@/components/gnb/HeaderBebridge";
+import HeaderBebridge from "@/components/gdummy/HeaderBebridge";
 import MemberList from "@/components/table/member/MemberList";
 // import SideMenu from "@/components/SideMenu/SideMenu";
 import InviteRecords from "@/components/table/InviteRecords";
